ru.po merged with msgmerge

+++ src/libromdata/disc/CIAReader.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (libromdata)                       *
  * CIAReader.cpp: Nintendo 3DS CIA reader.                                 *
  *                                                                         *
- * Copyright (c) 2016-2017 by David Korth.                                 *
+ * Copyright (c) 2016-2018 by David Korth.                                 *
  *                                                                         *
  * This program is free software; you can redistribute it and/or modify it *
  * under the terms of the GNU General Public License as published by the   *
@@ -14,9 +14,8 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           *
  * GNU General Public License for more details.                            *
  *                                                                         *
- * You should have received a copy of the GNU General Public License along *
- * with this program; if not, write to the Free Software Foundation, Inc., *
- * 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.           *
+ * You should have received a copy of the GNU General Public License       *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.   *
  ***************************************************************************/
 
 #include "librpbase/config.librpbase.h"
@@ -28,7 +27,6 @@
 #include "librpbase/crypto/AesCipherFactory.hpp"
 #include "librpbase/crypto/IAesCipher.hpp"
 #include "librpbase/crypto/KeyManager.hpp"
-#include "../crypto/CtrKeyScrambler.hpp"
 #include "../crypto/N3DSVerifyKeys.hpp"
 #endif /* ENABLE_DECRYPTION */
 using namespace LibRpBase;
@@ -368,8 +366,7 @@ size_t CIAReader::read(void *ptr, size_t size)
 		m_lastError = EIO;
 		return 0;
 	}
-	unsigned int sz_dec = d->cipher->decrypt(
-		static_cast<uint8_t*>(ptr), (unsigned int)size);
+	unsigned int sz_dec = d->cipher->decrypt(static_cast<uint8_t*>(ptr), size);
 	if (sz_dec != size) {
 		// decrypt() failed.
 		m_lastError = EIO;
